requirements.txt
cryptography>=41.0.0
argon2-cffi>=23.1.0
tqdm>=4.66.0